Implement /round1 server endpoint
This creates an endpoint at GET /round1 that returns a JSON object representing all the questions for the round.

Why do we do this instead of coding the questions into the frontend?  The answer is separation of concerns to prepare for ease of refactoring.  Right now, this is a hardcoded array of questions, but a future refactor will turn this into a database-backed set of questions, so we can store the questions for each round and let the endpoint focus only on retrieval and handoff to the frontend for display.