Fix infinite network loop
The useEffect runs every time the state changes (because we specified the listState and setListState functions in the useEffect.  This means we'll query the endpoint, get data, then set the state...which triggers the whole thing running again.

This adds a check to make sure it only queries the network when we have no questions